Bugs with FIXML, and cancelling works now. More coherent Cancelling interface
