+++ ally/FIXML/__init__.py
@@ -1,2 +1,2 @@
-from .FIXML import FIXML, OrderType
-from .		import utils
+from .methods	import FIXML, OrderType
+from .			import utils
