+++ ally/Order/__init__.py
@@ -1,4 +1,5 @@
 from .				import order
 from .Submit		import submit
-from .order			import Order
+# from .order			import Order
 from .Outstanding	import orders
+from .Directives	import *
